lab8-9/lab10/requirements.txt
Django==4.2.11
djangorestframework==3.14.0